Add 'enoughemail.com' label to the filtered emails
In a backwards compatible way, add the new required scope permission.
Then change the filter rule creation logic to also create if needed
and add 'enoughemail.com' label.

What's cool about this is that if you fetch an individual label, it will
have the count of emails tagged with it, making it possible to gauge the
stats of how many emails where filtered over time for each user