Allow creating provisional users
Allow creating user records without a uid i.e. if a user is not logged
in, creating a user with their email address will allow them to create
filter rules via email. Previously this would fail a signup, but now a
signup will work as this allows this provisional user to turn into a
real one